requirements.txt
python-dotenv==1.0.1
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
openai==1.40.2
gradio==4.39.0
matplotlib==3.9.0
pandas==2.2.2
dateparser==1.2.0
pydantic==2.8.2
vaderSentiment==3.3.2
pillow==10.4.0
